pyproject.toml
[project]
name = "aitaskmaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.80.12",
    "python-dotenv>=1.2.1",
]
